Alpha V3.2 (4/2/24)
The Revised Update: The live user count has been replaced with a more useful one that also includes many useful features, such as an FAQ. Also, the Airstore, which has been broken for a while, is now working once again. There are many other new changes and fixes as well, including a fix to the RSS app and the removal of the files app.